+++ packages/core/src/spawn.ts
@@ -23,8 +23,7 @@
 
 import { AnsiLogger, LogLevel, TimestampFormat } from 'node-ansi-logger';
 import { hasParameter } from '@matterbridge/utils';
-
-import { BroadcastServer } from './broadcastServer.js';
+import { BroadcastServer } from '@matterbridge/thread';
 
 /**
  * Spawns a child process with the given command and arguments.