+++ packages/thread/src/export.ts
@@ -1,2 +1,2 @@
 export * from './worker.js';
-export * from './workerTypes.js';
+export * from './broadcastServer.js';